LoveSimulation_plan/prototype/episode1/requirements.txt
Flask>=2.0
librosa>=0.10.0
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.10.0
praat-parselmouth>=0.4.0
anthropic>=0.7.0
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
